refactor: convert spawn builtin to package pattern
- Create openpaw/builtins/tools/spawn/ package with standard 4-file structure
  - __init__.py: SpawnToolBuiltin facade (185 lines, down from 595)
  - models.py: Pydantic input schemas (SpawnAgentInput, GetSubagentResultInput, CancelSubagentInput)
  - formatters.py: Shared formatting helpers (format_time_ago, format_duration, format_spawn_success)
  - tools.py: LangChain StructuredTool factory functions (5 tools)
- Delete monolithic openpaw/builtins/tools/spawn.py (595 lines)
- Add backward-compatible wrappers on SpawnToolBuiltin for _format_time_ago, _format_duration, _format_spawn_success
- All 2,834 tests pass, ruff clean, no behavioral changes
- Registry import path unchanged: openpaw.builtins.tools.spawn.SpawnToolBuiltin
